+++ crates/client/block_import/src/metrics.rs
@@ -1,7 +1,12 @@
+use mc_analytics::register_gauge_metric_instrument;
 use mc_db::MadaraBackend;
-use mc_metrics::{Gauge, MetricsRegistry, PrometheusError, F64};
 use mp_block::Header;
 use num_traits::FromPrimitive;
+use opentelemetry::metrics::Gauge;
+use opentelemetry::{
+    global::{self, Error},
+    KeyValue,
+};
 use std::{
     sync::Mutex,
     time::{Duration, Instant},
@@ -16,45 +21,108 @@ pub struct BlockMetrics {
     pub last_db_metrics_update_instant: Mutex<Option<Instant>>,
 
     // L2 network metrics
-    pub l2_block_number: Gauge<F64>,
-    pub l2_sync_time: Gauge<F64>,
-    pub l2_avg_sync_time: Gauge<F64>,
-    pub l2_latest_sync_time: Gauge<F64>,
-    pub l2_state_size: Gauge<F64>, // TODO: remove this, as well as the return value from db_metrics update.
-    pub transaction_count: Gauge<F64>,
-    pub event_count: Gauge<F64>,
+    pub l2_block_number: Gauge<u64>,
+    pub l2_sync_time: Gauge<f64>,
+    pub l2_avg_sync_time: Gauge<f64>,
+    pub l2_latest_sync_time: Gauge<f64>,
+    pub l2_state_size: Gauge<f64>, // TODO: remove this, as well as the return value from db_metrics update.
+    pub transaction_count: Gauge<u64>,
+    pub event_count: Gauge<u64>,
     // L1 network metrics
-    pub l1_gas_price_wei: Gauge<F64>,
-    pub l1_gas_price_strk: Gauge<F64>,
+    pub l1_gas_price_wei: Gauge<f64>,
+    pub l1_gas_price_strk: Gauge<f64>,
 }
 
 impl BlockMetrics {
-    pub fn register(starting_block: u64, registry: &MetricsRegistry) -> Result<Self, PrometheusError> {
+    pub fn register(starting_block: u64) -> Result<Self, Error> {
+        let common_scope_attributes = vec![KeyValue::new("crate", "block_import")];
+        let block_import_meter = global::meter_with_version(
+            "crates.block_import.opentelemetry",
+            Some("0.17"),
+            Some("https://opentelemetry.io/schemas/1.2.0"),
+            Some(common_scope_attributes.clone()),
+        );
+
+        let l2_block_number = register_gauge_metric_instrument(
+            &block_import_meter,
+            "l2_block_number".to_string(),
+            "Current block number".to_string(),
+            "".to_string(),
+        );
+
+        let l2_sync_time = register_gauge_metric_instrument(
+            &block_import_meter,
+            "l2_sync_time".to_string(),
+            "Complete sync time since startup in secs (does not account for restarts)".to_string(),
+            "".to_string(),
+        );
+
+        let l2_avg_sync_time = register_gauge_metric_instrument(
+            &block_import_meter,
+            "l2_avg_sync_time".to_string(),
+            "Average time spent between blocks since startup in secs".to_string(),
+            "".to_string(),
+        );
+
+        let l2_latest_sync_time = register_gauge_metric_instrument(
+            &block_import_meter,
+            "l2_latest_sync_time".to_string(),
+            "Latest time spent between blocks in secs".to_string(),
+            "".to_string(),
+        );
+
+        let l2_state_size = register_gauge_metric_instrument(
+            &block_import_meter,
+            "l2_state_size".to_string(),
+            "Node storage usage in GB".to_string(),
+            "".to_string(),
+        );
+
+        let transaction_count = register_gauge_metric_instrument(
+            &block_import_meter,
+            "transaction_count".to_string(),
+            "Latest block transaction count".to_string(),
+            "".to_string(),
+        );
+
+        let event_count = register_gauge_metric_instrument(
+            &block_import_meter,
+            "event_count".to_string(),
+            "Latest block event count".to_string(),
+            "".to_string(),
+        );
+
+        let l1_gas_price_wei = register_gauge_metric_instrument(
+            &block_import_meter,
+            "l1_gas_price_wei".to_string(),
+            "Latest block L1 ETH gas price".to_string(),
+            "".to_string(),
+        );
+
+        let l1_gas_price_strk = register_gauge_metric_instrument(
+            &block_import_meter,
+            "l1_gas_price_strk".to_string(),
+            "Latest block L1 STRK gas price".to_string(),
+            "".to_string(),
+        );
+
         Ok(Self {
             starting_block,
             starting_time: Instant::now(),
             last_update_instant: Default::default(),
             last_db_metrics_update_instant: Default::default(),
 
-            l2_block_number: registry.register(Gauge::new("madara_l2_block_number", "Current block number")?)?,
-            l2_sync_time: registry.register(Gauge::new(
-                "madara_l2_sync_time",
-                "Complete sync time since startup in secs (does not account for restarts)",
-            )?)?,
-            l2_avg_sync_time: registry.register(Gauge::new(
-                "madara_l2_avg_sync_time",
-                "Average time spent between blocks since startup in secs",
-            )?)?,
-            l2_latest_sync_time: registry
-                .register(Gauge::new("madara_l2_latest_sync_time", "Latest time spent between blocks in secs")?)?,
-            l2_state_size: registry.register(Gauge::new("madara_l2_state_size", "Node storage usage in GB")?)?,
-            transaction_count: registry
-                .register(Gauge::new("madara_transaction_count", "Latest block transaction count")?)?,
-            event_count: registry.register(Gauge::new("madara_event_count", "Latest block event count")?)?,
-            l1_gas_price_wei: registry
-                .register(Gauge::new("madara_l1_block_gas_price", "Latest block L1 ETH gas price")?)?,
-            l1_gas_price_strk: registry
-                .register(Gauge::new("madara_l1_block_gas_price_strk", "Latest block L1 STRK gas price")?)?,
+            l2_block_number,
+            l2_sync_time,
+            l2_avg_sync_time,
+            l2_latest_sync_time,
+            l2_state_size,
+
+            transaction_count,
+            event_count,
+
+            l1_gas_price_wei,
+            l1_gas_price_strk,
         })
     }
 
@@ -70,16 +138,17 @@ impl BlockMetrics {
         };
 
         let total_sync_time = now.duration_since(self.starting_time).as_secs_f64();
-        self.l2_sync_time.set(total_sync_time);
-        self.l2_latest_sync_time.set(latest_sync_time);
-        self.l2_avg_sync_time.set(total_sync_time / (block_header.block_number - self.starting_block) as f64);
 
-        self.l2_block_number.set(block_header.block_number as f64);
-        self.transaction_count.set(f64::from_u64(block_header.transaction_count).unwrap_or(0f64));
-        self.event_count.set(f64::from_u64(block_header.event_count).unwrap_or(0f64));
+        self.l2_sync_time.record(total_sync_time, &[]);
+        self.l2_latest_sync_time.record(latest_sync_time, &[]);
+        self.l2_avg_sync_time.record(total_sync_time / (block_header.block_number - self.starting_block) as f64, &[]);
 
-        self.l1_gas_price_wei.set(f64::from_u128(block_header.l1_gas_price.eth_l1_gas_price).unwrap_or(0f64));
-        self.l1_gas_price_strk.set(f64::from_u128(block_header.l1_gas_price.strk_l1_gas_price).unwrap_or(0f64));
+        self.l2_block_number.record(block_header.block_number, &[]);
+        self.transaction_count.record(block_header.transaction_count, &[]);
+        self.event_count.record(block_header.event_count, &[]);
+
+        self.l1_gas_price_wei.record(f64::from_u128(block_header.l1_gas_price.eth_l1_gas_price).unwrap_or(0f64), &[]);
+        self.l1_gas_price_strk.record(f64::from_u128(block_header.l1_gas_price.strk_l1_gas_price).unwrap_or(0f64), &[]);
 
         {
             let mut last_db_instant = self.last_db_metrics_update_instant.lock().expect("Poisoned lock");
@@ -89,7 +158,7 @@ impl BlockMetrics {
                 *last_db_instant = Some(now);
                 let storage_size = backend.update_metrics();
                 let size_gb = storage_size as f64 / (1024 * 1024 * 1024) as f64;
-                self.l2_state_size.set(size_gb);
+                self.l2_state_size.record(size_gb, &[]);
             }
         }
     }