+++ crates/client/block_import/src/verify_apply.rs
@@ -39,13 +39,13 @@ impl VerifyApply {
         block: PreValidatedBlock,
         validation: BlockValidationContext,
     ) -> Result<BlockImportResult, BlockImportError> {
-        log::debug!("acquiring verify_apply exclusive");
+        tracing::debug!("acquiring verify_apply exclusive");
         let _exclusive = self.mutex.lock().await;
-        log::debug!("acquired verify_apply exclusive");
+        tracing::debug!("acquired verify_apply exclusive");
 
         let backend = Arc::clone(&self.backend);
         let res = global_spawn_rayon_task(move || verify_apply_inner(&backend, block, validation)).await;
-        log::debug!("releasing verify_apply exclusive");
+        tracing::debug!("releasing verify_apply exclusive");
         res
     }
 
@@ -55,13 +55,13 @@ impl VerifyApply {
         block: PreValidatedPendingBlock,
         validation: BlockValidationContext,
     ) -> Result<PendingBlockImportResult, BlockImportError> {
-        log::debug!("acquiring verify_apply exclusive (pending)");
+        tracing::debug!("acquiring verify_apply exclusive (pending)");
         let _exclusive = self.mutex.lock().await;
-        log::debug!("acquired verify_apply exclusive (pending)");
+        tracing::debug!("acquired verify_apply exclusive (pending)");
 
         let backend = Arc::clone(&self.backend);
         let res = global_spawn_rayon_task(move || verify_apply_pending_inner(&backend, block, validation)).await;
-        log::debug!("releasing verify_apply exclusive (pending)");
+        tracing::debug!("releasing verify_apply exclusive (pending)");
         res
     }
 }
@@ -84,7 +84,7 @@ pub fn verify_apply_inner(
     // Block hash
     let (block_hash, header) = block_hash(&block, &validation, block_number, parent_block_hash, global_state_root)?;
 
-    log::debug!("verify_apply_inner store block {}", header.block_number);
+    tracing::debug!("verify_apply_inner store block {}", header.block_number);
 
     // store block, also uses rayon heavily internally
     backend
@@ -216,15 +216,15 @@ fn update_tries(
         return Ok(global_state_root);
     }
 
-    log::debug!(
+    tracing::debug!(
         "Deployed contracts: [{:?}]",
         block.state_diff.deployed_contracts.iter().map(|c| c.address.hex_display()).format(", ")
     );
-    log::debug!(
+    tracing::debug!(
         "Declared classes: [{:?}]",
         block.state_diff.declared_classes.iter().map(|c| c.class_hash.hex_display()).format(", ")
     );
-    log::debug!(
+    tracing::debug!(
         "Deprecated declared classes: [{:?}]",
         block.state_diff.deprecated_declared_classes.iter().map(|c| c.hex_display()).format(", ")
     );