+++ crates/client/gateway/src/server/handler.rs
@@ -1,18 +1,25 @@
 use std::sync::Arc;
 
-use hyper::{body, Body, Request, Response};
+use bytes::Buf;
+use http_body_util::BodyExt;
+use hyper::{body::Incoming, Request, Response};
 use mc_db::MadaraBackend;
-use mc_rpc::providers::AddTransactionProvider;
+use mc_rpc::{
+    providers::AddTransactionProvider,
+    versions::v0_7_1::methods::trace::trace_block_transactions::trace_block_transactions as v0_7_1_trace_block_transactions,
+    Starknet,
+};
 use mp_block::{BlockId, BlockTag, MadaraBlock, MadaraMaybePendingBlockInfo, MadaraPendingBlock};
 use mp_class::{ClassInfo, ContractClass};
 use mp_gateway::{
     block::{BlockStatus, ProviderBlock, ProviderBlockPending, ProviderBlockSignature},
     state_update::{ProviderStateUpdate, ProviderStateUpdatePending},
 };
+use serde::Serialize;
 use serde_json::json;
 use starknet_core::types::{
     BroadcastedDeclareTransaction, BroadcastedDeployAccountTransaction, BroadcastedInvokeTransaction,
-    BroadcastedTransaction,
+    BroadcastedTransaction, TransactionTraceWithHash,
 };
 use starknet_types_core::felt::Felt;
 
@@ -26,7 +33,10 @@ use super::{
     },
 };
 
-pub async fn handle_get_block(req: Request<Body>, backend: Arc<MadaraBackend>) -> Result<Response<Body>, GatewayError> {
+pub async fn handle_get_block(
+    req: Request<Incoming>,
+    backend: Arc<MadaraBackend>,
+) -> Result<Response<String>, GatewayError> {
     let params = get_params_from_request(&req);
     let block_id = block_id_from_params(&params).or_internal_server_error("Retrieving block id")?;
 
@@ -80,9 +90,9 @@ pub async fn handle_get_block(req: Request<Body>, backend: Arc<MadaraBackend>) -
 }
 
 pub async fn handle_get_signature(
-    req: Request<Body>,
+    req: Request<Incoming>,
     backend: Arc<MadaraBackend>,
-) -> Result<Response<Body>, GatewayError> {
+) -> Result<Response<String>, GatewayError> {
     let params = get_params_from_request(&req);
     let block_id = block_id_from_params(&params).or_internal_server_error("Retrieving block id")?;
 
@@ -112,9 +122,9 @@ pub async fn handle_get_signature(
 }
 
 pub async fn handle_get_state_update(
-    req: Request<Body>,
+    req: Request<Incoming>,
     backend: Arc<MadaraBackend>,
-) -> Result<Response<Body>, GatewayError> {
+) -> Result<Response<String>, GatewayError> {
     let params = get_params_from_request(&req);
     let block_id = block_id_from_params(&params).or_internal_server_error("Retrieving block id")?;
 
@@ -216,10 +226,30 @@ pub async fn handle_get_state_update(
     }
 }
 
-pub async fn handle_get_class_by_hash(
-    req: Request<Body>,
+pub async fn handle_get_block_traces(
+    req: Request<Incoming>,
     backend: Arc<MadaraBackend>,
-) -> Result<Response<Body>, GatewayError> {
+    add_transaction_provider: Arc<dyn AddTransactionProvider>,
+) -> Result<Response<String>, GatewayError> {
+    let params = get_params_from_request(&req);
+    let block_id = block_id_from_params(&params).or_internal_server_error("Retrieving block id")?;
+
+    #[derive(Serialize)]
+    struct BlockTraces {
+        traces: Vec<TransactionTraceWithHash>,
+    }
+
+    let traces =
+        v0_7_1_trace_block_transactions(&Starknet::new(backend, add_transaction_provider), block_id.into()).await?;
+    let block_traces = BlockTraces { traces };
+
+    Ok(create_json_response(hyper::StatusCode::OK, &block_traces))
+}
+
+pub async fn handle_get_class_by_hash(
+    req: Request<Incoming>,
+    backend: Arc<MadaraBackend>,
+) -> Result<Response<String>, GatewayError> {
     let params = get_params_from_request(&req);
     let block_id = block_id_from_params(&params).unwrap_or(BlockId::Tag(BlockTag::Latest));
 
@@ -240,7 +270,7 @@ pub async fn handle_get_class_by_hash(
                 .as_ref()
                 .serialize_to_json()
                 .or_internal_server_error("Failed to serialize legacy class")?;
-            create_response_with_json_body(hyper::StatusCode::OK, &class)
+            create_response_with_json_body(hyper::StatusCode::OK, class)
         }
     };
 
@@ -248,9 +278,9 @@ pub async fn handle_get_class_by_hash(
 }
 
 pub async fn handle_get_compiled_class_by_class_hash(
-    req: Request<Body>,
+    req: Request<Incoming>,
     backend: Arc<MadaraBackend>,
-) -> Result<Response<Body>, GatewayError> {
+) -> Result<Response<String>, GatewayError> {
     let params = get_params_from_request(&req);
     let block_id = block_id_from_params(&params).unwrap_or(BlockId::Tag(BlockTag::Latest));
 
@@ -274,10 +304,10 @@ pub async fn handle_get_compiled_class_by_class_hash(
         .or_internal_server_error(format!("Retrieving compiled Sierra class from class hash {class_hash:x}"))?
         .ok_or(StarknetError::class_not_found(class_hash))?;
 
-    Ok(create_response_with_json_body(hyper::StatusCode::OK, class_compiled.as_ref()))
+    Ok(create_response_with_json_body(hyper::StatusCode::OK, class_compiled.0))
 }
 
-pub async fn handle_get_contract_addresses(backend: Arc<MadaraBackend>) -> Result<Response<Body>, GatewayError> {
+pub async fn handle_get_contract_addresses(backend: Arc<MadaraBackend>) -> Result<Response<String>, GatewayError> {
     let chain_config = &backend.chain_config();
     Ok(create_json_response(
         hyper::StatusCode::OK,
@@ -288,18 +318,18 @@ pub async fn handle_get_contract_addresses(backend: Arc<MadaraBackend>) -> Resul
     ))
 }
 
-pub async fn handle_get_public_key(backend: Arc<MadaraBackend>) -> Result<Response<Body>, GatewayError> {
+pub async fn handle_get_public_key(backend: Arc<MadaraBackend>) -> Result<Response<String>, GatewayError> {
     let public_key = &backend.chain_config().private_key.public;
     Ok(create_string_response(hyper::StatusCode::OK, format!("\"{:#x}\"", public_key)))
 }
 
 pub async fn handle_add_transaction(
-    req: Request<Body>,
+    req: Request<Incoming>,
     add_transaction_provider: Arc<dyn AddTransactionProvider>,
-) -> Result<Response<Body>, GatewayError> {
-    let whole_body = body::to_bytes(req.into_body()).await.or_internal_server_error("Failed to read request body")?;
+) -> Result<Response<String>, GatewayError> {
+    let whole_body = req.collect().await.or_internal_server_error("Failed to read request body")?.aggregate();
 
-    let transaction = serde_json::from_slice::<BroadcastedTransaction>(whole_body.as_ref())
+    let transaction = serde_json::from_reader::<_, BroadcastedTransaction>(whole_body.reader())
         .map_err(|e| GatewayError::StarknetError(StarknetError::malformed_request(e)))?;
 
     let response = match transaction {
@@ -314,7 +344,7 @@ pub async fn handle_add_transaction(
 async fn declare_transaction(
     tx: BroadcastedDeclareTransaction,
     add_transaction_provider: Arc<dyn AddTransactionProvider>,
-) -> Response<Body> {
+) -> Response<String> {
     match add_transaction_provider.add_declare_transaction(tx).await {
         Ok(result) => create_json_response(hyper::StatusCode::OK, &result),
         Err(e) => create_json_response(hyper::StatusCode::OK, &e),
@@ -324,7 +354,7 @@ async fn declare_transaction(
 async fn deploy_account_transaction(
     tx: BroadcastedDeployAccountTransaction,
     add_transaction_provider: Arc<dyn AddTransactionProvider>,
-) -> Response<Body> {
+) -> Response<String> {
     match add_transaction_provider.add_deploy_account_transaction(tx).await {
         Ok(result) => create_json_response(hyper::StatusCode::OK, &result),
         Err(e) => create_json_response(hyper::StatusCode::OK, &e),
@@ -334,7 +364,7 @@ async fn deploy_account_transaction(
 async fn invoke_transaction(
     tx: BroadcastedInvokeTransaction,
     add_transaction_provider: Arc<dyn AddTransactionProvider>,
-) -> Response<Body> {
+) -> Response<String> {
     match add_transaction_provider.add_invoke_transaction(tx).await {
         Ok(result) => create_json_response(hyper::StatusCode::OK, &result),
         Err(e) => create_json_response(hyper::StatusCode::OK, &e),