+++ crates/client/mempool/src/inner.rs
@@ -95,7 +95,6 @@ pub struct NonceChain {
     transactions: BTreeMap<OrderMempoolTransactionByNonce, ()>,
     front_arrived_at: ArrivedAtTimestamp,
     front_nonce: Nonce,
-    #[cfg(debug_assertions)]
     front_tx_hash: TransactionHash,
 }
 
@@ -121,7 +120,6 @@ impl NonceChain {
     pub fn new_with_first_tx(tx: MempoolTransaction) -> Self {
         Self {
             front_arrived_at: tx.arrived_at,
-            #[cfg(debug_assertions)]
             front_tx_hash: tx.tx_hash(),
             front_nonce: tx.nonce(),
             transactions: iter::once((OrderMempoolTransactionByNonce(tx), ())).collect(),
@@ -146,7 +144,6 @@ impl NonceChain {
     ) -> Result<(InsertedPosition, ReplacedState), TxInsersionError> {
         let mempool_tx_arrived_at = mempool_tx.arrived_at;
         let mempool_tx_nonce = mempool_tx.nonce();
-        #[cfg(debug_assertions)]
         let mempool_tx_hash = mempool_tx.tx_hash();
 
         let replaced = if force {
@@ -159,7 +156,6 @@ impl NonceChain {
             match self.transactions.entry(OrderMempoolTransactionByNonce(mempool_tx)) {
                 btree_map::Entry::Occupied(entry) => {
                     // duplicate nonce, either it's because the hash is duplicated or nonce conflict with another tx.
-                    #[cfg(debug_assertions)]
                     if entry.key().0.tx_hash() == mempool_tx_hash {
                         return Err(TxInsersionError::DuplicateTxn);
                     } else {
@@ -176,10 +172,7 @@ impl NonceChain {
             // We insrted at the front here
             let former_head_arrived_at = core::mem::replace(&mut self.front_arrived_at, mempool_tx_arrived_at);
             self.front_nonce = mempool_tx_nonce;
-            #[cfg(debug_assertions)]
-            {
-                self.front_tx_hash = mempool_tx_hash;
-            }
+            self.front_tx_hash = mempool_tx_hash;
             InsertedPosition::Front { former_head_arrived_at }
         } else {
             InsertedPosition::Other
@@ -587,7 +580,7 @@ mod tests {
     struct MempoolInvariantsProblem(Vec<Operation>);
     impl MempoolInvariantsProblem {
         fn check(&self) {
-            log::debug!("\n\n\n\n\nCase: {:#?}", self);
+            tracing::debug!("\n\n\n\n\nCase: {:#?}", self);
             let mut mempool = MempoolInner::default();
             mempool.check_invariants();
 
@@ -613,14 +606,14 @@ mod tests {
                 } else {
                     assert!(inserted.is_empty())
                 }
-                log::trace!("Popped {:?}", res.map(|el| Insert(el, false)));
+                tracing::trace!("Popped {:?}", res.map(|el| Insert(el, false)));
             };
 
             for op in &self.0 {
                 match op {
                     Operation::Insert(insert) => {
                         let force = insert.1;
-                        log::trace!("Insert {:?}", insert);
+                        tracing::trace!("Insert {:?}", insert);
                         let res = mempool.insert_tx(insert.0.clone(), insert.1);
 
                         let expected = if !force
@@ -645,20 +638,20 @@ mod tests {
                             inserted_contract_nonce_pairs.insert((insert.0.nonce(), insert.0.contract_address()));
                         }
 
-                        log::trace!("Result {:?}", res);
+                        tracing::trace!("Result {:?}", res);
                     }
                     Operation::Pop => {
-                        log::trace!("Pop");
+                        tracing::trace!("Pop");
                         let res = mempool.pop_next();
                         handle_pop(res, &mut inserted, &mut inserted_contract_nonce_pairs, &mut new_contracts);
                     }
                 }
-                log::trace!("State: {mempool:#?}");
+                tracing::trace!("State: {mempool:#?}");
                 mempool.check_invariants();
             }
 
             loop {
-                log::trace!("Pop");
+                tracing::trace!("Pop");
                 let Some(res) = mempool.pop_next() else { break };
                 handle_pop(Some(res), &mut inserted, &mut inserted_contract_nonce_pairs, &mut new_contracts);
                 mempool.check_invariants();
@@ -666,7 +659,7 @@ mod tests {
             assert!(inserted.is_empty());
             assert!(inserted_contract_nonce_pairs.is_empty());
             assert!(new_contracts.is_empty());
-            log::trace!("Done :)");
+            tracing::trace!("Done :)");
         }
     }
 
@@ -675,7 +668,7 @@ mod tests {
         #[test]
         fn proptest_mempool(pb in any::<MempoolInvariantsProblem>()) {
             let _ = env_logger::builder().is_test(true).try_init();
-            log::set_max_level(log::LevelFilter::Trace);
+            tracing::log::set_max_level(tracing::log::LevelFilter::Trace);
             pb.check();
         }
     }