+++ crates/node/src/service/rpc.rs
@@ -4,9 +4,7 @@ use jsonrpsee::server::ServerHandle;
 use tokio::task::JoinSet;
 
 use mc_db::DatabaseService;
-use mc_metrics::MetricsRegistry;
 use mc_rpc::{providers::AddTransactionProvider, versioned_rpc_api, Starknet};
-use mp_chain_config::ChainConfig;
 use mp_utils::service::Service;
 
 use metrics::RpcMetrics;
@@ -26,8 +24,6 @@ impl RpcService {
     pub fn new(
         config: &RpcParams,
         db: &DatabaseService,
-        chain_config: Arc<ChainConfig>,
-        metrics_handle: &MetricsRegistry,
         add_txs_method_provider: Arc<dyn AddTransactionProvider>,
     ) -> anyhow::Result<Self> {
         if config.rpc_disabled {
@@ -39,7 +35,7 @@ impl RpcService {
             (RpcMethods::Unsafe, _) => (true, true),
             (RpcMethods::Auto, false) => (true, true),
             (RpcMethods::Auto, true) => {
-                log::warn!(
+                tracing::warn!(
                     "Option `--rpc-external` will hide node operator endpoints. To enable them, please pass \
                      `--rpc-methods unsafe`."
                 );
@@ -47,8 +43,8 @@ impl RpcService {
             }
         };
         let (read, write, trace, internal, ws) = (rpcs, rpcs, rpcs, node_operator, rpcs);
-        let starknet = Starknet::new(Arc::clone(db.backend()), chain_config.clone(), add_txs_method_provider);
-        let metrics = RpcMetrics::register(metrics_handle)?;
+        let starknet = Starknet::new(Arc::clone(db.backend()), add_txs_method_provider);
+        let metrics = RpcMetrics::register()?;
 
         Ok(Self {
             server_config: Some(ServerConfig {